+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `Server::getOnlinePlayers` for getting all online players.
 - `Player::kick` for kicking a specific player with a custom message.
+- `Player::getPing` for getting the average latency of a connected player.
 
 ### Changed